+++ deploy.sh
@@ -35,8 +35,6 @@ echo "Place multi-demo's files."
 cp -r ./projects/demo/dist/* ./$directory/demo/.
 echo "Place Substrate-connect extension's zip."
 cp ./projects/extension/dist/packed-extension.zip ./$directory/extension/packed-extension.zip
-echo "Place landing page's files."
-cp -r ./projects/landing-page/dist/* ./$directory/.
 echo "Generate API docs."
 yarn api-docs