+++ projects/extension/src/components/NetworkTab.tsx
@@ -3,7 +3,7 @@ import { Accordion } from "."
 
 import { NetworkTabProps, App, OptionsNetworkTabHealthContent } from "../types"
 import "../main.css"
-import IconWeb3 from "./IconWeb3"
+import { IconWeb3 } from "./IconWeb3"
 
 interface NetworkContentProps {
   health?: OptionsNetworkTabHealthContent
@@ -41,7 +41,7 @@ const NetworkContent = ({ network, health, apps }: NetworkContentProps) => {
       </div>
       <div className="flex flex-row py-2">
         <div className="basis-1/5 font-bold">Apps</div>
-        <div className="basis-4/5">{apps.length}:</div>
+        <div className="basis-4/5">{apps.length}</div>
       </div>
       <div className="flex flex-row pb-2">
         <div className="basis-1/5"></div>
@@ -66,7 +66,7 @@ const NetworkTab: FunctionComponent<NetworkTabProps> = ({
     <NetworkContent health={health} apps={apps} network={name} />,
   ]
   return (
-    <div className="flex w-full max-w-2xl mb-3 items-baseline">
+    <div className="flex w-full mb-3 items-baseline font-roboto">
       <Accordion
         titles={[
           <div className="flex rounded-lg">
@@ -74,11 +74,14 @@ const NetworkTab: FunctionComponent<NetworkTabProps> = ({
               <IconWeb3>{name.toLowerCase()}</IconWeb3>
               <div className="txt-xl cap">
                 {name}
-                <span className="pl-2 text-[#616161]">({apps.length})</span>
+                <span className="pl-2 text-[#616161]">
+                  {apps.length ? "(" + apps.length + ")" : ""}
+                </span>
               </div>
             </div>
           </div>,
         ]}
+        origin={"options"}
         contents={contents}
         showTitleIcon={!!contents.length}
       />